// submodules/common-dto/api-client/main/models/ArtworkOrderStatus.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Status of order
 */
export enum ArtworkOrderStatus {
    PENDING = 'pending',
    PAID = 'paid',
    DELIVERING = 'delivering',
    PAYOUTS_PROCEED = 'payouts_proceed',
    COMPLETED = 'completed',
    CANCELED = 'canceled',
}